+++ wayland-protocols/src/stable.rs
@@ -14,6 +14,21 @@ pub mod presentation_time {
     );
 }
 
+pub mod xdg_shell {
+    //! XDG Shell protocol
+    //!
+    //! Exposes the `xdg_wm_base` global, which deprecates and replace `wl_shell`.
+
+    wayland_protocol!(
+        "xdg-shell",
+        [
+            (wl_seat, wl_seat_interface),
+            (wl_surface, wl_surface_interface),
+            (wl_output, wl_output_interface)
+        ]
+    );
+}
+
 pub mod viewporter {
     //! Viewporter protocol
     //!